Struct xor_name::Prefix

pub struct Prefix { /* private fields */ }

A section prefix, i.e. a sequence of bits specifying the part of the network’s name space consisting of all names that start with this sequence.

Implementations

impl Prefix

pub fn new(bit_count: usize, name: XorName) -> Self

Creates a new Prefix with the first bit_count bits of name. Insignificant bits are all set to 0.

pub fn name(&self) -> XorName

Returns the name of this prefix.

pub fn pushed(self, bit: bool) -> Self

Returns self with an appended bit: 0 if bit is false, and 1 if bit is true. If self.bit_count is already at the maximum for this type, then an unmodified copy of self is returned.

pub fn popped(self) -> Self

Returns a prefix copying the first bitcount() - 1 bits from self, or self if it is already empty.

pub fn bit_count(&self) -> usize

Returns the number of bits in the prefix.

pub fn is_empty(&self) -> bool

Returns true if this is the empty prefix, with no bits.

pub fn is_compatible(&self, other: &Self) -> bool

Returns true if self is a prefix of other or vice versa.

pub fn is_extension_of(&self, other: &Self) -> bool

Returns true if other is compatible but strictly shorter than self.

pub fn is_neighbour(&self, other: &Self) -> bool

Returns true if the other prefix differs in exactly one bit from this one.

pub fn common_prefix(&self, name: &XorName) -> usize

Returns the number of common leading bits with the input name, capped with prefix length.

pub fn matches(&self, name: &XorName) -> bool

Returns true if this is a prefix of the given name.

pub fn cmp_distance(&self, other: &Self, target: &XorName) -> Ordering

Compares the distance of self and other to target. Returns Less if self is closer, Greater if other is closer, and compares the prefix directly if of equal distance (this is to make sorting deterministic).

pub fn cmp_breadth_first(&self, other: &Self) -> Ordering

Compares the prefixes using breadth-first order. That is, shorter prefixes are ordered before longer. This is in contrast with the default Ord impl of Prefix which uses depth-first order.

pub fn lower_bound(&self) -> XorName

Returns the smallest name matching the prefix

pub fn upper_bound(&self) -> XorName

Returns the largest name matching the prefix

pub fn range_inclusive(&self) -> RangeInclusive<XorName>

Inclusive range from lower_bound to upper_bound

pub fn is_covered_by<'a, I>(&self, prefixes: I) -> bool where
I: IntoIterator<Item = &'a Self> + Clone,

Returns whether the namespace defined by self is covered by prefixes in the prefixes set

pub fn with_flipped_bit(&self, i: u8) -> Self

Returns the neighbouring prefix differing in the i-th bit If i is larger than our bit count, self is returned

pub fn substituted_in(&self, name: XorName) -> XorName

Returns the given name with first bits replaced by self

pub fn sibling(&self) -> Self

Returns the same prefix, with the last bit flipped, or unchanged, if empty.

pub fn ancestor(&self, bit_count: u8) -> Self

Returns the ancestors of this prefix that has the given bit count.

Panics

Panics if bit_count is not less than the bit count of this prefix.

pub fn ancestors(&self) -> Ancestors

Returns an iterator that yields all ancestors of this prefix.