Struct xor_name::XorName

pub struct XorName(pub [u8; XOR_NAME_LEN]);

A XOR_NAME_BITS-bit number, viewed as a point in XOR space.

This has as its only field an array of XOR_NAME_LEN bytes, i. e. a number between 0 and 2XOR_NAME_BITS - 1, the XorName's "ID".

XOR space is the space of these numbers, with the XOR metric as a notion of distance, i. e. the points with IDs x and y are considered to have distance x xor y.

Methods

impl XorName

fn new(id: [u8; XOR_NAME_LEN]) -> XorName

Construct a XorName from a XOR_NAME_LEN byte array.

fn get_id(&self) -> [u8; XOR_NAME_LEN]

Return the internal array.

fn as_hex(&self) -> String

Hex-encode the XorName as a String.

fn bucket_distance(&self, name: &XorName) -> usize

Deprecated

Currently identical to bucket_index. This method will be replaced with XOR_NAME_BITS - bucket_index or removed entirely.

fn bucket_index(&self, name: &XorName) -> usize

Returns the number of leading bits in which self and name agree.

Here, "leading bits" means the most significant bits. E. g. for 10101... and 10011..., that value will be 2, as their common prefix 10 has length 2 and the third bit is the first one in which they disagree.

Equivalently, this is XOR_NAME_BITS - bucket_distance, where bucket_distance is the length of the remainders after the common prefix is removed from the IDs of self and name.

The bucket distance is the magnitude of the XOR distance. More precisely, if d > 0 is the XOR distance between self and name, the bucket distance equals floor(log2(d)), i. e. a bucket distance of n means that 2n - 1 <= d < 2n.

fn cmp_closeness(&self, lhs: &XorName, rhs: &XorName) -> Ordering

Compares lhs and rhs with respect to their distance from self.

fn from_hex(s: &str) -> Result<XorName, XorNameFromHexError>

Hex-decode a XorName from a &str.