Crate xor_name

Structs

BitIndexOutOfBoundsError
A bit was indexed by a value i that did not satisfy 0 <= i < XOR_NAME_BITS.

XorName
A XOR_NAME_BITS-bit number, viewed as a point in XOR space.

Enums

XorNameFromHexError
Errors that can occur when decoding a XorName from a string.

Constants

XOR_NAME_BITS
Constant bit length of XorName.

XOR_NAME_LEN
Constant byte length of XorName.

Functions

closer_to_target
Returns true if lhs is closer to target than rhs.

closer_to_target_or_equal
Returns true if lhs is closer to target than rhs, or when lhs == rhs.

slice_as_u8_64_array

slice_equal
Returns true if both slices are equal in length and have equal contents.