Crate xor_name

xor_name

TODO requires further documentation.

Re-exports

pub use rand;

Macros

xor_name

Creates XorName with the given leading bytes and the rest filled with zeroes.

Structs

Prefix

A section prefix, i.e. a sequence of bits specifying the part of the network’s name space consisting of all names that start with this sequence.

XorName

A 256-bit number, viewed as a point in XOR space.

Constants

XOR_NAME_LEN

Constant byte length of XorName.